ADD seed.rb
`data/development.sqlite.backup` has been converted as seed ruby file.

A few things to note about that :

* backup database was outdated and missing a lot of migrations
* data contained was not valid anymore

Data was fixed to fullfill validations. As it was mostly fk pseudo
constraints, it can be considered quite safe, even if we don't get the
whole codebase for now.

Seeding database is fully destructive, wipping all data from database.
This ensures we can force records id. As such, it should only be run as
a setup step (and well, it's all it's about, after all).

Fix #3